Create ui util, remove onStateChange
Creted ui.js util and moved relevant functions into it.
Removed any references to now-removed stateChange handler.
Small cleanup changes (unused stat import, inconsistent styling).
